Refactoring classes and methods for i18n
